+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # stylelint-config-ultimate 更新日志
 
-## version 0.1.0  开启 stylelint-config-ultimate 之旅  (2016-10-27)
+## version 0.1.0  开启 stylelint-config-ultimate 之旅  (2018-9-11)
 
 * 初始化 stylelint-config-ultimate
-* 基于 stylelint v9.3.0、 stylelint-config-standard v18.2.0 和 stylelint-order v0.8.1 定制
+* 基于 stylelint v9.5.0、 stylelint-config-standard v18.2.0 和 stylelint-order v1.0.0 定制